gentoo resync : 03.11.2021

+++ sys-libs/glibc/glibc-2.33-r7.ebuild
@@ -23,7 +23,7 @@ PATCH_DEV=dilfridge
 if [[ ${PV} == 9999* ]]; then
 	inherit git-r3
 else
-	KEYWORDS="~alpha ~amd64 ~arm ~arm64 ~hppa ~ia64 ~m68k ~mips ~ppc ~ppc64 ~riscv ~s390 ~sparc ~x86"
+	KEYWORDS="~alpha amd64 arm arm64 ~hppa ~ia64 ~m68k ~mips ppc ppc64 ~riscv ~s390 ~sparc x86"
 	SRC_URI="mirror://gnu/glibc/${P}.tar.xz"
 	SRC_URI+=" https://dev.gentoo.org/~${PATCH_DEV}/distfiles/${P}-patches-${PATCH_VER}.tar.xz"
 fi
@@ -1506,7 +1506,12 @@ pkg_preinst() {
 	if ! use crypt && has_version "${CATEGORY}/${PN}[crypt]"; then
 		PRESERVED_OLD_LIBCRYPT=1
 		preserve_old_lib /$(get_libdir)/libcrypt$(get_libname 1)
-		cp "${EROOT}"/usr/include/crypt.h "${T}"/crypt.h || die
+
+		# Only copy if it exists; some people may have tiny embedded
+		# systems without headers: https://bugs.gentoo.org/802207#c16.
+		if [[ -f "${EROOT}"/usr/include/crypt.h ]] ; then
+			cp "${EROOT}"/usr/include/crypt.h "${T}"/crypt.h || die
+		fi
 	else
 		PRESERVED_OLD_LIBCRYPT=0
 	fi
@@ -1542,10 +1547,17 @@ pkg_postinst() {
 
 	if [[ ${PRESERVED_OLD_LIBCRYPT} -eq 1 ]] ; then
 		preserve_old_lib_notify /$(get_libdir)/libcrypt$(get_libname 1)
-		cp "${T}"/crypt.h "${EROOT}"/usr/include/crypt.h || eerror "Error restoring crypt.h, please file a bug"
+
+		# Only copy if it exists; some people may have tiny embedded
+		# systems without headers: https://bugs.gentoo.org/802207#c16
+		if [[ -f "${T}"/crypt.h ]] ; then
+			cp "${T}"/crypt.h "${EROOT}"/usr/include/crypt.h || eerror "Error restoring crypt.h, please file a bug"
+		fi
+
 		elog "Please ignore a possible later error message about a file collision involving"
 		elog "/usr/include/crypt.h. We need to preserve this file for the moment to keep"
 		elog "the upgrade working, but it also needs to be overwritten when"
 		elog "sys-libs/libxcrypt is installed. See bug 802210 for more details."
+		elog "If you have FEATURES=collision-protect, please use FEATURES=unmerge-orphans instead!"
 	fi
 }
